# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos y Órdenes",
        "description": "Documentación  de la API, incluyendo los endpoints de gestión de productos y órdenes.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local de desarrollo"
        },
        {
            "url": "https://api.example.com",
            "description": "Servidor de producción"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener todos los productos",
                "description": "Este endpoint devuelve una lista de todos los productos registrados. Si no hay productos, retorna un mensaje indicando que no se encontraron registros.",
                "operationId": "1adde327047214eb0921fdba340afbda",
                "responses": {
                    "200": {
                        "description": "Lista de productos o mensaje de que no hay productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "description": "Lista de productos",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Producto A"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Electrónica"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descripción del producto"
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "example": 10
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "No hay productos registrados"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un producto",
                "description": "Este endpoint permite crear un nuevo producto. Requiere los campos 'name', 'category', 'description' y 'amount'. Devuelve el producto creado o un error en caso de fallo en la validación o creación.",
                "operationId": "919825ff5882341c963c4ae575e8eb9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Producto A"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Electrónica"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción del producto"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Producto A"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "Electrónica"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Descripción del producto"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "example": 10
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la validación de datos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validación de datos"
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo name es obligatorio."
                                                    }
                                                },
                                                "category": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo category es obligatorio."
                                                    }
                                                },
                                                "description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo description es obligatorio."
                                                    }
                                                },
                                                "amount": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "El campo amount debe ser un número."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear el producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error al crear el producto"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}